DADOS DA OPERAÇAO DIVISAO

+++ treina.c
@@ -1,23 +1,23 @@
 #include <stdio.h>
  
 int main (){
-    int nota1, nota2, nota3;
-    int media;
+    float nota1, nota2, nota3;
+    float media;
 
     printf("*** programa de calculo de media ***\n ");
 
     printf("Digite a sua primeira nota\n ");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
 
     printf("Digite a sua segunda  nota\n ");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
 
     printf("Digite a sua  terceira nota\n ");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
     media =(nota1 + nota2 + nota3) / 3;
 
-    printf("A media e : %d", media);
+    printf("A media e : %.2f", media);
